test: add frontend tests

+++ frontend/src/app/module/sailor/pages/delete-sailor/delete-sailor.component.ts
@@ -1,4 +1,4 @@
-import {Component} from '@angular/core';
+import {Component, EventEmitter, Output} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {MessageService} from '@data/service/message.service';
 import {MatSnackBar} from '@angular/material/snack-bar';
@@ -9,8 +9,9 @@ import {DeleteSailorCommand} from '@model/domain/message/command/delete-sailor-c
   styleUrls: ['./delete-sailor.component.scss'],
 })
 export class DeleteSailorComponent {
-  public wasSubmitted = false;
+  public wasDeleted = false;
   public form: FormGroup;
+  @Output() public sailorDeleted = new EventEmitter<void>();
 
   public constructor(
     private messageService: MessageService,
@@ -24,12 +25,18 @@ export class DeleteSailorComponent {
 
   public commandSubmit(): void
   {
+    if (!this.form.valid) {
+      return;
+    }
+
     const command: DeleteSailorCommand = {
       email: this.form.get('email')?.value,
     };
+
     this.messageService.deleteSailor(command).subscribe(() => {
-      this.wasSubmitted = true;
+      this.wasDeleted = true;
       this.snackBar.open('Sailor removed successfully', '',{duration: 5000});
+      this.sailorDeleted.emit();
     });
   }
 }
